+++ SocialMedia.Application/Services/PostRepository.cs
@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Http;
 using SocialMedia.Application.Mapper;
 using SocialMedia.Infrastructure.Models;
 using Microsoft.EntityFrameworkCore;
+using SocialMedia.Application.Response;
 
 
 namespace SocialMedia.Application.Repository
@@ -18,7 +19,7 @@ namespace SocialMedia.Application.Repository
 	{
 		public PostRepository() {}
 		AppDbContext _context = new AppDbContext();
-		public async Task<int> AddAsync(dtoPost NewPost)
+		public async Task<Response<int>> AddAsync(dtoPost NewPost)
 		{
 			Post post = NewPost.ToPost();
 			try
@@ -27,10 +28,9 @@ namespace SocialMedia.Application.Repository
 				await _context.SaveChangesAsync();
 			}catch(Exception ex)
 			{
-				throw new Exception(ex.Message);
+				return Response<int>.Failure("faild to add post");
 			}
-			return post.Id;
-
+			return Response<int>.Success(post.Id,"",200);
 		}
 
 		public async Task<Post?> Find(int id)
@@ -38,10 +38,27 @@ namespace SocialMedia.Application.Repository
 			return await _context.posts.FirstOrDefaultAsync(p=>p.Id == id);
 		}                               
 
-		public async void AddPostImage(Post post , string ImageName)
+		public async Task<Response<string>> AddPostImage(IHostingEnvironment _host,IFormFile image,int id)
 		{
+			
+			Post? post = await Find(id);
+			
+			if(post == null)return Response<string>.Failure("Post Not Found");
+			
+			string ImageName;
+			//add post photo
+			ImageName = AddPostPhoto(_host, image, "postPhotos");
+			//add user photo
+			ImageName = AddPostPhoto(_host, image, "userPhotos");
+	
 			post.ImagePath = ImageName;
-			await _context.SaveChangesAsync();
+			try{
+				await _context.SaveChangesAsync();
+			}catch(Exception ex)
+			{
+				return Response<string>.Failure("faild to add post image");
+			}
+			return Response<string>.Success("Image Post Added.");
 		}
 		public string AddPostPhoto(IHostingEnvironment _host, IFormFile image , string path)
 		{
@@ -53,46 +70,39 @@ namespace SocialMedia.Application.Repository
 			return ImageName;
 		}
 
-		public async Task<List<Post>> GetAll()
+		public async Task<Response<List<Post>>> GetAll()
 		{
-			List<Post> posts = new List<Post>(); 
-			try
-			{
-				posts = await _context.posts.OrderByDescending(item => item.Id).ToListAsync();
-			}catch(Exception ex)
-			{
-				throw new Exception(ex.Message);
-			}
-			return posts;
+			return Response<List<Post>>.Success(await _context.posts.OrderByDescending(item => item.Id).ToListAsync());
 		}
 
-		public async Task<List<Post>> GetAllByUser(string userId)
+		public async Task<Response<List<Post>>> GetAllByUser(string userId)
 		{
-			List<Post> posts = new List<Post>(); 
-			try
-			{
-				posts = await _context.posts.Where((item) => item.UserId == userId).OrderByDescending((item) => item.Id).ToListAsync();
-			}catch(Exception ex)
-			{
-				throw new Exception(ex.Message);
-			}
-			return posts;
+			
+			return Response<List<Post>>.Success(await _context.posts.Where((item) => item.UserId == userId).OrderByDescending((item) => item.Id).ToListAsync());
 		}
 
-		public async Task Delete(Post post)
+		public async Task<Response<string>> Delete(int id)
 		{
+			Post? post = await Find(id);
+			if(post == null)
+			{
+				return Response<string>.Failure("Post Not Found.");
+			}
 			try
 			{
 				_context.posts.Remove(post);
+				
 				await _context.SaveChangesAsync();
 			}catch(Exception ex)
 			{	
-				throw new Exception(ex.Message);
+				return Response<string>.Failure("Faild to delete post.");
 			}
+			return Response<string>.Success("Post Deleted.");
 		}
 
-		public async Task UpdateReact(Post post , string type , int payload)
+		public async Task UpdateReact(int postid , string type , int payload)
 		{
+			Post? post = await Find(postid);
 			if (type == "like")
 			{
 				post.likes += payload;
@@ -126,8 +136,13 @@ namespace SocialMedia.Application.Repository
 			}
 		}
 
-		public async Task Update(string content , string status , Post post)
+		public async Task<Response<string>> Update(string content , string status , int id)
 		{
+			Post? post = await Find(id);
+			if (post == null)
+			{
+				return Response<string>.Failure("Post Not Found.");
+			}
 			try
 			{
 				post.Content = content;
@@ -135,8 +150,9 @@ namespace SocialMedia.Application.Repository
 				await _context.SaveChangesAsync();
 			}catch(Exception ex)
 			{
-				throw new Exception(ex.Message);
+				return Response<string>.Failure("Faild to update post.");
 			}
+			return Response<string>.Success("Post updated.");
 		}
 		 
 		public async Task<List<Post>> GetFollowingPost(string id)
